+++ src/Plugin/Behat.php
@@ -14,7 +14,15 @@ use PHPCensor\Plugin;
  */
 class Behat extends Plugin
 {
+
+  /**
+    * @var string
+    */
     protected $features;
+
+    /**
+     * @var string
+     */
     protected $executable;
 
     /**
@@ -31,14 +39,14 @@ class Behat extends Plugin
     public function __construct(Builder $builder, Build $build, array $options = [])
     {
         parent::__construct($builder, $build, $options);
+        
+        if (isset($options['executable'])) {
+          $this->executable = $options['executable'];
+        } else {
+          $this->executable = $this->findBinary('behat');
+        }
 
         $this->features = '';
-
-        if (isset($options['executable'])) {
-            $this->executable = $options['executable'];
-        } else {
-            $this->executable = $this->findBinary('behat');
-        }
 
         if (!empty($options['features'])) {
             $this->features = $options['features'];
@@ -53,15 +61,13 @@ class Behat extends Plugin
         $currentDir = getcwd();
         chdir($this->builder->buildPath);
 
-        $behat = $this->executable;
-
-        if (!$behat) {
+        if (!$this->executable) {
             $this->builder->logFailure(sprintf('Could not find %s', 'behat'));
 
             return false;
         }
 
-        $success = $this->builder->executeCommand($behat . ' %s', $this->features);
+        $success = $this->builder->executeCommand($this->executable . ' %s', $this->features);
         chdir($currentDir);
 
         list($errorCount, $data) = $this->parseBehatOutput();