+++ src/Plugin/SensiolabsInsight.php
@@ -33,7 +33,7 @@ class SensiolabsInsight extends Plugin
      * @var int
      */
     protected $allowedWarnings;
-
+    protected $executable;
     /**
      * @return string
      */
@@ -65,6 +65,13 @@ class SensiolabsInsight extends Plugin
         if (array_key_exists('project_uuid', $options)) {
             $this->projectUuid = $options['project_uuid'];
         }
+        
+        if (array_key_exists('executable', $options)) {
+            $this->executable = $options['executable'];
+        } else {
+            $this->executable = $this->findBinary('insight');
+        }
+
     }
 
     /**
@@ -74,7 +81,7 @@ class SensiolabsInsight extends Plugin
      */
     public function execute()
     {
-        $insightBinaryPath = $this->findBinary('insight');
+        $insightBinaryPath = $this->executable;
 
         $this->executeSensiolabsInsight($insightBinaryPath);